requirements.txt
yfinance
pandas
python-dotenv
google-genai
